CreateGrid function and Style changes
Finished the createGrid function which makes the grid based on the selected size (currently users cannot change this by interacting with the website)

made some visual changes, removed the border , added a background, changed background color for buttons, made the buttons change color while hovering.

Made the color of the squares in the grid change color based on whether the mouse is hovering over it. (currently does not maintain color after the mouse is off a particular square).