Random Color button
Added the color button again to be able to switch between random colors and the user selected color modes.

Instead of using a function, used variables to state which mode was active. Used if statements to check which mode was active. Added event listeners that switch states.

Used a random generated number and converted to hex code to be able to select new random color.

Once button click is active, changes the background of the button to show the button click occurred.